Major refactor
I wasn't able to get comms working, and I didn't know why, and I was
frustrated by the tight coupling that ESPHome-RATGDO had with its
Security+ 2.0 comms code and the inscrutability of the replay-based sync
in the mqtt codebase, so I threw it all out and started over.

Anyway, it works, though the homekit stuff is still a bit wonky and I've
disabled improv so that I can mash buttons manually.