Fix fof occasional spurious door unlock notifications
wait for a status command to be processes to properly set the initial state of
all homekit characteristics.  Also timeout if we don't receive a status in
a reasonable amount of time.  This prevents unintentional state changes if
a home hub reads the state before we initialize everything
Note, secplus1 doesnt have a status command so it will just timeout

This specifically addresses an issue where we get spurious door unlock
notifications, but will also address incorrect door states immediately
after booting.